requirements.txt
# requirements.txt
Flask
google-generativeai
python-dotenv